requirements.txt
phue==1.1
pipewire_python==0.2.3
systemd_python==235
